src/requirements.txt
django
djangorestframework
djangorestframework-jwt
django-filter
django-crispy-forms
coverage
django-nose
ipdbplugin  # for ipdb support on test failures
ipdb
freezegun  # to mock datetime.now
